Completed 1st refactor of SoilAPI module and comment out plot_cpt_fence method in SoilPlot class for future implementation
